requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
idna==3.3
mercantile==1.2.1
numpy==1.22.1
python-dotenv==0.19.2
ratelimit==2.2.1
requests==2.27.1
urllib3==1.26.8
